Update log.cpp
New "createTextFile" function to store creds external to Ps1.
Updated "log::Sendlog" for the updated PowerShell script changes. Option to have as without ZIP of the log.txt file.
Updated "createlog" to have it check number of log files in folder and delete when hitting specified number and then refresh.